+++ packages/expo-blur/src/BlurView.types.ts
@@ -10,6 +10,8 @@ export type BlurViewProps = {
    * A number from `1` to `100` to control the intensity of the blur effect.
    *
    * You can animate this property using `Animated API` from React Native or using `react-native-reanimated`.
+   * > If using `react-native-reanimated` see [`animatedProps`](#animatedprops) for more details.
+   * >
    * > Animating this property using `Animated API` from React Native with `setNativeDriver: true` does not work.
    *
    * @default 50
@@ -26,6 +28,14 @@ export type BlurViewProps = {
    *
    */
   blurReductionFactor?: number;
+  /**
+   * Animated props used for animating the BlurView's intensity.
+   * This prop has to be created with `Animated.createAnimatedProps` from `react-native-reanimated`.
+   *
+   * > Note: When animatedProps are passed the underlying blur component will automatically be marked as Animated, therefore it is not
+   * necessary to create an animated version of the `BlurView`.
+   */
+  animatedProps?: { intensity?: number };
 } & ViewProps;
 
 export type BlurTint = 'light' | 'dark' | 'default';